+++ src/main/java/kitchenpos/application/OrderService.java
@@ -10,7 +10,7 @@ import kitchenpos.dao.OrderTableDao;
 import kitchenpos.domain.order.Order;
 import kitchenpos.domain.order.OrderLineItem;
 import kitchenpos.domain.order.OrderStatus;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.table.OrderTable;
 import kitchenpos.repository.OrderRepository;
 
 import org.springframework.stereotype.Service;
