+++ src/main/java/kitchenpos/application/dto/convertor/TableConvertor.java
@@ -5,7 +5,7 @@ import java.util.stream.Collectors;
 
 import kitchenpos.application.dto.request.OrderTableRequest;
 import kitchenpos.application.dto.response.OrderTableResponse;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.table.OrderTable;
 
 public class TableConvertor {
