+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -3,7 +3,6 @@ package kitchenpos.application;
 import static kitchenpos.domain.fixture.MenuFixture.후라이드_치킨_세트의_가격과_메뉴_상품_리스트는;
 import static kitchenpos.domain.fixture.MenuGroupFixture.치킨_세트;
 import static kitchenpos.domain.fixture.MenuProductFixture.상품_하나;
-import static kitchenpos.domain.fixture.OrderFixture.요리중인_주문;
 import static kitchenpos.domain.fixture.OrderFixture.주문_1번의_주문_항목들은;
 import static kitchenpos.domain.fixture.OrderTableFixture.비어있는_테이블;
 import static kitchenpos.domain.fixture.ProductFixture.후라이드_치킨;
@@ -33,12 +32,12 @@ import kitchenpos.dao.fake.FakeOrderLineItemDao;
 import kitchenpos.dao.fake.FakeOrderTableDao;
 import kitchenpos.dao.fake.FakeProductDao;
 import kitchenpos.domain.menu.Menu;
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.domain.menu.MenuGroup;
 import kitchenpos.domain.menu.MenuProduct;
 import kitchenpos.domain.order.Order;
 import kitchenpos.domain.order.OrderLineItem;
 import kitchenpos.domain.order.OrderStatus;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.table.OrderTable;
 import kitchenpos.domain.Product;
 import kitchenpos.domain.fixture.OrderTableFixture;
 import kitchenpos.repository.OrderRepository;
