+++ src/test/java/kitchenpos/domain/PriceTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain.product;
+package kitchenpos.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
