+++ src/test/java/kitchenpos/domain/table/OrderTableTest.java
@@ -1,9 +1,11 @@
-package kitchenpos.domain;
+package kitchenpos.domain.table;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import kitchenpos.domain.table.OrderTable;
 
 @DisplayName("Order 도메인 테스트")
 class OrderTableTest {
